db-api/requirements.txt
flask
google-cloud-secret-manager
google-cloud-firestore
google-api-core
requests
debugpy
bcrypt